Split Mode into parts
* Mode class simplified. Now it was only following methods

  __init__ for creating Modes (typically not for end users)
  __call__ for setting some settings for the Mode. Not needed in very
           basic usage (for power users)
  __enter__ for context manager protocol
  __exit__ for context manager protocol

  and these attributes:

  .methods (list of Method classes to be used)
  .manager (the activation manager)

* The Mode has now only these responsibilities:
  1) Offer easy way of defining Modes as ordered collections of methods
  2) Provide the main wakepy API in the form on context managers

* Split dbus adapter resolution from Mode to calls.py. This should be
  responsibility of the future Caller class.

* Create modeactivator.py and move the ModeWorkerThread there. In the
  future, there should be a ModeActivator class here which does the
  heavy lifting and activates/deactivates Modes using Methods.

  TODO: Create the ModeActivator class

* Split ModeActivationManager out from the Mode. This is also WIP and
  not functional / tests.

  TODO: Finish and make tests for ModeActivationManager class.
